Add content resizing while merging page
Add to pageObject methods to merge a transformed page. These methods are
similar to mergePage, but a transformation matrix is applied to the merged
page.

It is possible to give the transformation matrix by hand (cf. § 8.3.3
"Common Transformations" of the PDF reference), using mergeTransformedPage,
but helper methods also provide common operations such as scaling,
resizing, translating and some of their combinaisons.

An use case for these methode is to put two A4 portrait pages reduced to
A5 on an A4 landscape page. This is the important feature that was missing
to be able to create booklets with pyPdf.

Patch from Kjo Hansi Glaz <kjo@a4nancy.net.eu.org>