Do not change pages given to PageObject.mergeTransformedPage
When using PageObject.mergeTransformedPage(self, page2), the
transformation matrix was directly applied by changing the content
stream in page2. As a result, if page2 was used afterward, it was
modified by the previously given transformation matrix.

To fix this, we now modify the content stream from page2 immediately
before merging it, by supplying a transformation as an argument of
a new PageObject._mergePage private method.

Patch by Kjo Hansi Glaz